GASTELO

+++ app/Http/Controllers/ConceptoController.php
@@ -3,46 +3,61 @@
 namespace App\Http\Controllers;
 
 use Validator;
-use Illuminate\Http\Request;
 use App\Http\Requests;
+use App\Persona;
+use App\Caja;
+use App\Acciones;
+use App\Ahorros;
+use App\Credito;
+use App\Cuota;
 use App\Concepto;
+use App\Transaccion;
+use App\ControlPersona;
 use App\Librerias\Libreria;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 
 class ConceptoController extends Controller
 {
-
     protected $folderview      = 'app.concepto';
-    protected $tituloAdmin     = 'Concepto';
+    protected $tituloAdmin     = 'Conceptos';
     protected $tituloRegistrar = 'Registrar concepto';
     protected $tituloModificar = 'Modificar concepto';
     protected $tituloEliminar  = 'Eliminar concepto';
     protected $rutas           = array('create' => 'concepto.create', 
-            'edit'     => 'concepto.edit', 
-            'delete'   => 'concepto.eliminar',
-            'search'   => 'concepto.buscar',
-            'index'    => 'concepto.index',
+            'edit'   => 'concepto.edit', 
+            'delete' => 'concepto.eliminar',
+            'search' => 'concepto.buscar',
+            'index'  => 'concepto.index',
         );
 
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
     public function __construct()
     {
         $this->middleware('auth');
     }
 
+    /**
+     * Mostrar el resultado de búsquedas
+     * 
+     * @return Response 
+     */
     public function buscar(Request $request)
     {
         $pagina           = $request->input('page');
         $filas            = $request->input('filas');
         $entidad          = 'Concepto';
-        $concepto         = Libreria::getParam($request->input('concepto'));
         $tipo             = Libreria::getParam($request->input('tipo'));
-        $resultado        = Concepto::listar($concepto, $tipo);
+        $resultado        = Concepto::listar($tipo);
         $lista            = $resultado->get();
         $cabecera         = array();
         $cabecera[]       = array('valor' => '#', 'numero' => '1');
-        $cabecera[]       = array('valor' => 'Concepto', 'numero' => '1');
+        $cabecera[]       = array('valor' => 'Titulo', 'numero' => '1');
         $cabecera[]       = array('valor' => 'Tipo', 'numero' => '1');
         $cabecera[]       = array('valor' => 'Operaciones', 'numero' => '2');
         
@@ -73,8 +88,9 @@ class ConceptoController extends Controller
         $entidad          = 'Concepto';
         $title            = $this->tituloAdmin;
         $titulo_registrar = $this->tituloRegistrar;
+        $cboTipo        = [''=>'Todo']+ array('I'=>'Ingresos','E'=>'Egresos');
         $ruta             = $this->rutas;
-        return view($this->folderview.'.admin')->with(compact('entidad', 'title', 'titulo_registrar', 'ruta'));
+        return view($this->folderview.'.admin')->with(compact('entidad','cboTipo' ,'title', 'titulo_registrar', 'ruta'));
     }
 
     /**
@@ -84,13 +100,14 @@ class ConceptoController extends Controller
      */
     public function create(Request $request)
     {
-        $listar       = Libreria::getParam($request->input('listar'), 'NO');
-        $entidad      = 'Concepto';
-        $concepto  = null;
-        $formData     = array('concepto.store');
-        $formData     = array('route' => $formData, 'class' => 'form-horizontal', 'id' => 'formMantenimiento'.$entidad, 'autocomplete' => 'off');
-        $boton        = 'Registrar'; 
-        return view($this->folderview.'.mant')->with(compact('concepto', 'formData', 'entidad', 'boton', 'listar'));
+        $listar         = Libreria::getParam($request->input('listar'), 'NO');
+        $entidad        = 'Concepto';
+        $concepto        = null;
+        $formData       = array('concepto.store');
+        $cboTipo        = [''=>'Seleccione']+ array('I'=>'Ingresos','E'=>'Egresos');
+        $formData       = array('route' => $formData, 'class' => 'form-horizontal', 'id' => 'formMantenimiento'.$entidad, 'autocomplete' => 'off');
+        $boton          = 'Registrar'; 
+        return view($this->folderview.'.mant')->with(compact('concepto', 'cboTipo','formData', 'entidad', 'boton', 'listar'));
     }
 
     /**
@@ -102,16 +119,18 @@ class ConceptoController extends Controller
     public function store(Request $request)
     {
         $listar     = Libreria::getParam($request->input('listar'), 'NO');
-        $reglas     = array('concepto' => 'required|max:120');
-        $mensajes   = array();
-        $validacion = Validator::make($request->all(), $reglas, $mensajes);
+        $reglas = array(
+            'titulo'         => 'required',
+            'tipo'         => 'required'
+            );
+        $validacion = Validator::make($request->all(),$reglas);
         if ($validacion->fails()) {
             return $validacion->messages()->toJson();
         }
         $error = DB::transaction(function() use($request){
-            $concepto       = new Concepto();
-            $concepto->concepto = strtoupper($request->input('concepto'));
-            $concepto->tipo = $request->input('tipo');
+            $concepto               = new Concepto();
+            $concepto->titulo        = $request->input('titulo');
+            $concepto->tipo        = $request->input('tipo');
             $concepto->save();
         });
         return is_null($error) ? "OK" : $error;
@@ -134,19 +153,20 @@ class ConceptoController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Request $request, $id)
+    public function edit($id, Request $request)
     {
         $existe = Libreria::verificarExistencia($id, 'concepto');
         if ($existe !== true) {
             return $existe;
         }
-        $listar   = Libreria::getParam($request->input('listar'), 'NO');
-        $concepto = Concepto::find($id);
-        $entidad  = 'Concepto';
-        $formData = array('concepto.update', $id);
-        $formData = array('route' => $formData, 'method' => 'PUT', 'class' => 'form-horizontal', 'id' => 'formMantenimiento'.$entidad, 'autocomplete' => 'off');
-        $boton    = 'Modificar';
-        return view($this->folderview.'.mant')->with(compact('concepto', 'formData', 'entidad', 'boton', 'listar'));
+        $listar         = Libreria::getParam($request->input('listar'), 'NO');
+        $concepto        = Concepto::find($id);
+        $entidad        = 'Concepto';
+        $formData       = array('concepto.update', $id);
+        $cboTipo        = [''=>'Seleccione']+ array('I'=>'Ingresos','E'=>'Egresos');
+        $formData       = array('route' => $formData, 'method' => 'PUT', 'class' => 'form-horizontal', 'id' => 'formMantenimiento'.$entidad, 'autocomplete' => 'off');
+        $boton          = 'Modificar';
+        return view($this->folderview.'.mant')->with(compact('concepto','cboTipo', 'formData', 'entidad', 'boton', 'listar'));
     }
 
     /**
@@ -162,16 +182,18 @@ class ConceptoController extends Controller
         if ($existe !== true) {
             return $existe;
         }
-        $reglas     = array('concepto' => 'required|max:100');
-        $mensajes   = array();
-        $validacion = Validator::make($request->all(), $reglas, $mensajes);
+        $reglas = array(
+            'titulo'       => 'required|max:20|unique:concepto,titulo,'.$id.',id,deleted_at,NULL',
+            'tipo'         => 'required'
+            );
+        $validacion = Validator::make($request->all(),$reglas);
         if ($validacion->fails()) {
             return $validacion->messages()->toJson();
         } 
         $error = DB::transaction(function() use($request, $id){
-            $concepto       = Concepto::find($id);
-            $concepto->concepto = strtoupper($request->input('concepto'));
-            $concepto->tipo = $request->input('tipo');
+            $concepto                 = Concepto::find($id);
+            $concepto->titulo        = $request->input('titulo');
+            $concepto->tipo        = $request->input('tipo');
             $concepto->save();
         });
         return is_null($error) ? "OK" : $error;
@@ -212,10 +234,18 @@ class ConceptoController extends Controller
         if (!is_null(Libreria::obtenerParametro($listarLuego))) {
             $listar = $listarLuego;
         }
+        $count_acciones = Acciones::where('concepto_id',$id)->count();
+        $count_transaccion = Transaccion::where('concepto_id',$id)->count();
+
         $modelo   = Concepto::find($id);
         $entidad  = 'Concepto';
-        $formData = array('route' => array('concepto.destroy', $id), 'method' => 'DELETE', 'class' => 'form-horizontal', 'id' => 'formMantenimiento'.$entidad, 'autocomplete' => 'off');
         $boton    = 'Eliminar';
-        return view('app.confirmarEliminar')->with(compact('modelo', 'formData', 'entidad', 'boton', 'listar'));
+        if(($count_acciones==0) && ($count_transaccion == 0)){
+            $formData = array('route' => array('concepto.destroy', $id), 'method' => 'DELETE', 'class' => 'form-horizontal', 'id' => 'formMantenimiento'.$entidad, 'autocomplete' => 'off');
+            return view('app.confirmarEliminar')->with(compact('modelo', 'formData', 'entidad', 'boton', 'listar'));
+        }else{
+            return view($this->folderview.'.messageconcepto')->with(compact('modelo', 'formData', 'entidad', 'boton', 'listar'));
+        }
+        
     }
 }
